Defend against transient Puppet constant error
Observed once on site, the second time this monkey patch was loaded it seemed to
cause an error related to constant Puppet when loading this file.

This seems to be a load-order type error, and as such is difficult to reproduce.
I was unable to reproduce the issue when I tried with a trivial PE 2015.3.3
master and a Windows 2012r2 vagrant VM.

The only thing I can think of at the moment is that Puppet.warn_once and other
logging methods are being called before they're actually loaded.  As a result
I'm putting a require 'puppet' at the top to try and make sure the logging
methods are available when the monkey patch is auto-loaded.